pyproject.toml
[tool.poetry]
name = "easy_notes_api"
version = "0.1.0"
description = ""
authors = ["Egor Sennov <esennov2005@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.3"
pydantic-settings = "^2.3.4"
isort = "^5.13.2"
pylint = "^3.2.5"
black = "^24.4.2"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
aiohttp = "^3.9.5"
url-normalize = "^1.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
